backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.12
redis==5.2.0
celery==5.4.0
minio==7.2.10
httpx==0.27.2
email-validator==2.2.0
bleach==6.2.0
pgvector==0.3.6
orjson==3.10.7
tenacity==9.0.0
slowapi==0.1.9
websockets==13.1
pytest==8.3.3
pytest-asyncio==0.24.0
prometheus-client==0.21.1

